// backend/src/modules/solicitudes/domain/enums/estado-solicitud.enum.ts
export enum EstadoSolicitud {
    DRAFT = 'draft',
    SUBMITTED = 'submitted',
    IN_REVIEW = 'in_review',
    PENDING_VALIDATION = 'pending_validation',
    APPROVED = 'approved',
    REJECTED = 'rejected',
    ABANDONED = 'abandoned',
}

export enum NivelRiesgo {
    BAJO = 'BAJO',
    MEDIO = 'MEDIO',
    ALTO = 'ALTO',
}